pyproject.toml
[tool.poetry]
name = "faust"
version = "0.1.0"
description = ""
authors = ["aiden <aiden@safuventures.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Faker = "^13.4.0"
environs = "^9.5.0"
confluent-kafka = "^1.8.2"
faust-streaming = "^0.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
